+++ packages/nadle/src/core/handlers/index.ts
@@ -2,6 +2,7 @@ import { ListHandler } from "./list-handler.js";
 import { GraphHandler } from "./graph-handler.js";
 import { WatchHandler } from "./watch-handler.js";
 import { DryRunHandler } from "./dry-run-handler.js";
+import { ExplainHandler } from "./explain-handler.js";
 import { ExecuteHandler } from "./execute-handler.js";
 import { CleanCacheHandler } from "./clean-cache-handler.js";
 import { ShowConfigHandler } from "./show-config-handler.js";
@@ -13,6 +14,7 @@ export const Handlers: HandlerConstructor[] = [
 	ListWorkspacesHandler,
 	CleanCacheHandler,
 	GraphHandler,
+	ExplainHandler,
 	DryRunHandler,
 	ShowConfigHandler,
 	WatchHandler,